+++ Teste/Fraude Votação/bad-user-1.py
@@ -100,11 +100,12 @@ def verifySignal(message, sig, public_key):
         return 0
 
 def main():  
+    IP = str(input("Entre com o Broke IP: "))
     qtd_usuarios = int(input("Informe a quantidade de usuarios: "))
     usuarios, chaves, eleitos, votacao = [], [], [], []
     
-    random.seed(random.randint(0,2^(32)-1))
-    nodeID = random.randint(0,2^(32)-1)
+    random.seed(random.randint(0,2^(32-1)))
+    nodeID = random.randint(0,2^(32-1))
     
     numero = str(nodeID)
     
@@ -127,14 +128,14 @@ def main():
                 dic = {"NodeId": nodeID, "PubKey": public_key.read()}
                 public_key.close()
                 
-                jsonSTR = json.dumps(dic,indent=2)
+                jsonSTR = json.dumps(dic)
                 
                 channel.basic_publish(exchange = 'ppd/pubkey', routing_key = '', body = jsonSTR)
                 print(usuarios)
 
     def callback1(ch, method, properties, body):
         try:
-            if(usuarios.index(json.dumps({"NodeId": nodeID},indent=2)) >= 0):
+            if(usuarios.index(json.dumps({"NodeId": nodeID})) >= 0):
                 pass
         except:
             sys.exit(0)
@@ -154,11 +155,11 @@ def main():
                 for i in range(qtd_usuarios-1, -1, -1):
                     user = json.loads(usuarios[i]) 
                     dic = {"NodeId":user["NodeId"],"ElectionNumber":nodeID}
-                    jsonSTR = json.dumps(dic,indent=2)
+                    jsonSTR = json.dumps(dic)
                     sig = genereteSignal(jsonSTR)
                     
                     dic.update({"Sign":sig})
-                    jsonSTR = json.dumps(dic,indent=2)
+                    jsonSTR = json.dumps(dic)
                     channel.basic_publish(exchange = 'ppd/election', routing_key = '', body = jsonSTR)
             elif(dic["NodeId"] == nodeID):
                 channel.basic_publish(exchange = 'ppd/init', routing_key = '', body = temp)
@@ -186,7 +187,7 @@ def main():
                 for chave in chaves:
                     chave = json.loads(chave)
                     if(chave["NodeId"] == dic["NodeId"]):
-                        if(not verifySignal(json.dumps(dic,indent=2), sig, chave["PubKey"])):
+                        if(not verifySignal(json.dumps(dic), sig, chave["PubKey"])):
                             print("\nLog: \n\t Tentativa de fraude na eleição")
                         else:
                             eleitos.append(temp)
@@ -205,12 +206,12 @@ def main():
                     challenger      = getChallenge(trasactionID)
                     
                     dic = {"NodeId":nodeID, "TransactionNumber":int(getTransactionID()),"Challenge":int(challenger)}
-                    jsonSTR = json.dumps(dic, indent=2)
+                    jsonSTR = json.dumps(dic)
                     
                     sig = genereteSignal(jsonSTR)
                     
                     dic.update({"Sign":sig})
-                    jsonSTR = json.dumps(dic,indent=2)
+                    jsonSTR = json.dumps(dic)
                     
                     channel.basic_publish(exchange = 'ppd/challenge', routing_key = '', body = jsonSTR)
                 
@@ -260,7 +261,7 @@ def main():
         for chave in chaves:
             chave = json.loads(chave)
             if(chave["NodeId"] == chairman["NodeId"]):
-                if(not verifySignal(json.dumps(dic,indent=2), sig, chave["PubKey"])):
+                if(not verifySignal(json.dumps(dic), sig, chave["PubKey"])):
                     print("\nLog: \n\t Tentativa de fraude na chave do lider")
                     return
         
@@ -271,7 +272,7 @@ def main():
         flag = True
         seed, multThread = [], []
 
-        for i in range(1,challenger*2+1):
+        for i in range(1,19):
             thread = threading.Thread(target=getSeed, args=(challenger, seed, i, ))
             multThread.append(thread)
             thread.start()
@@ -292,11 +293,11 @@ def main():
         
         #enviar resposta para broker
         dic = {"NodeId":nodeID, "TransactionNumber":int(getTransactionID()), "Seed":seed[0]}
-        jsonSTR = json.dumps(dic, indent=2)
+        jsonSTR = json.dumps(dic)
         sig = genereteSignal(jsonSTR)
         
         dic.update({"Sign":sig})
-        jsonSTR = json.dumps(dic,indent=0)
+        jsonSTR = json.dumps(dic)
         
         channel.basic_publish(exchange = 'ppd/solution', routing_key = '', body = jsonSTR)
         
@@ -332,7 +333,7 @@ def main():
         for chave in chaves:
             chave = json.loads(chave)
             if(chave["NodeId"] == dic["NodeId"]):
-                if(not verifySignal(json.dumps(dic,indent=2), sig, chave["PubKey"])):
+                if(not verifySignal(json.dumps(dic), sig, chave["PubKey"])):
                     print("\nLog: \n\t Tentativa de Fraude na Seed")
                     return
         
@@ -352,11 +353,11 @@ def main():
             arq.close()
             
             dic = {"NodeId":nodeID, "SolutionID":dic["NodeId"], "TransactionNumber":int(getTransactionID()), "Seed":dic["Seed"], "Vote":voto}
-            jsonSTR = json.dumps(dic, indent=2)
+            jsonSTR = json.dumps(dic)
             sig = genereteSignal(jsonSTR)
             
             dic.update({"Sign":sig})
-            jsonSTR = json.dumps(dic,indent=2)
+            jsonSTR = json.dumps(dic)
         
             channel.basic_publish(exchange = 'ppd/voting', routing_key = '', body = jsonSTR)  
         
@@ -388,7 +389,7 @@ def main():
                 for chave in chaves:
                     chave = json.loads(chave)
                     if(chave["NodeId"] == dic["NodeId"]):
-                        if(not verifySignal(json.dumps(dic,indent=2), sig, chave["PubKey"])):
+                        if(not verifySignal(json.dumps(dic), sig, chave["PubKey"])):
                             print("\nLog: \n\t Tentativa de Fraude na Votação")
                         else:
                             votacao.append(temp)
@@ -412,11 +413,11 @@ def main():
                     voto = json.loads(random.choice(usuarios))
                     
                     dic = {"NodeId":nodeID,"ElectionNumber":voto["NodeId"]}                    
-                    jsonSTR = json.dumps(dic,indent=2)
+                    jsonSTR = json.dumps(dic)
                     sig = genereteSignal(jsonSTR)
                     
                     dic.update({"Sign":sig})
-                    jsonSTR = json.dumps(dic,indent=2)
+                    jsonSTR = json.dumps(dic)
 
                     channel.basic_publish(exchange = 'ppd/election', routing_key = '', body = jsonSTR)
                     
@@ -426,7 +427,12 @@ def main():
                     arq.write("/".join(split[1:]))
                     arq.close()
                             
-    connection = pika.BlockingConnection(pika.ConnectionParameters(host = 'localhost'))
+    try:
+        credentials = pika.credentials.PlainCredentials("admin", "admin")
+        connection = pika.BlockingConnection(pika.ConnectionParameters(host = IP, credentials=credentials))
+    except:
+        connection = pika.BlockingConnection(pika.ConnectionParameters(host = 'localhost'))
+
     channel = connection.channel()
 
     print(nodeID)
@@ -459,7 +465,7 @@ def main():
     
     # InitMsg
     dic = {"NodeId": nodeID}
-    jsonSTR = json.dumps(dic,indent=2)
+    jsonSTR = json.dumps(dic)
     
     channel.basic_consume(queue = 'ppd/init/'+numero , on_message_callback = callback, auto_ack = True)
     channel.basic_publish(exchange = 'ppd/init', routing_key = '', body = jsonSTR)
@@ -487,9 +493,13 @@ def main():
 
 if __name__ == '__main__':
     try:
-        file = 'output/banco-de-dados.csv'
-        if(os.path.exists(file) and os.path.isfile(file)): 
-            os.remove(file)
+        file = './output'
+        if(os.path.exists(file)): 
+            file = file+ '/banco-de-dados.csv'
+            if(os.path.exists(file) and os.path.isfile(file)): 
+                os.remove(file)
+        else:
+            os.mkdir('./output')
         file = 'chaves/private_key.pem'
         if(os.path.exists(file) and os.path.isfile(file)): 
             os.remove(file)
